+++ RELEASE_NOTES.md
@@ -3,7 +3,7 @@
 ## v0.8.0 (2025-11-xx)
 
 - `catalog`
-  - Datbase migrations enabled
+  - Database migrations enabled
     - Migration 1: supporting future migrations and normalizing `ctime`
   - `getFacets` now returns a sorted list
   - Fix for tracks being duplicated in queue when multiple matching